refactor(tests): switch to snapshot testing and move tests into test folder
